Move WGL public functions to WGLUtil
